DockerCompose/backend/requirements.txt
# Backend Python dependencies for containerized deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.7

# PyTorch with CUDA 12.1 support
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.4.1

# MONAI and ML
monai>=1.4.0
numpy>=1.24,<3.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
matplotlib>=3.6.3
